Refactor: Refactored the listings page to make it a server component for faster INP, LCP and FCP
