+++ app/page.tsx
@@ -1,21 +1,13 @@
-"use client";
-
-import { useState, useEffect, useRef, Suspense } from "react";
-import React from "react";
+import { Suspense } from "react";
 import Link from "next/link";
-import { Search, MapPin } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Card, CardContent } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
 import dynamic from "next/dynamic";
 import { CategoriesSkeleton } from "@/components/categories-skeleton";
-import { getRecentListings, DisplayListingItem } from "@/lib/data";
-import { useRouter } from "next/navigation";
-import { useSuspenseQuery, useInfiniteQuery } from "@tanstack/react-query";
+import { getRecentListings } from "@/lib/data";
 import { ErrorBoundary } from "react-error-boundary";
 import { RecentListingsSkeleton } from "@/components/skeletons/recent-listings-skeleton";
-import Image from "next/image";
+import { RecentListings } from "@/components/recent-listings";
+import { SearchBar } from "@/components/search-bar";
+import { Button } from "@/components/ui/button"; // Added Button import
 
 const CategoriesSection = dynamic(
   () => import("../components/categories-section"),
@@ -29,133 +21,9 @@ const CategoriesSection = dynamic(
   },
 );
 
-function RecentListings() {
-  const {
-    data,
-    error,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-    status,
-  } = useInfiniteQuery({
-    queryKey: ["recentListings"],
-    queryFn: ({ pageParam = 1 }) => getRecentListings(pageParam),
-    getNextPageParam: (lastPage, pages) => {
-      return lastPage.length === 20 ? pages.length + 1 : undefined;
-    },
-    initialPageParam: 1,
-  });
-
-  const observerRef = useRef<IntersectionObserver | null>(null);
-  const loadingRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
-          fetchNextPage();
-        }
-      },
-      { threshold: 1.0 },
-    );
-
-    const currentLoadingRef = loadingRef.current; // Capture the current value
-
-    if (currentLoadingRef) {
-      observer.observe(currentLoadingRef);
-    }
-
-    observerRef.current = observer;
-
-    return () => {
-      if (currentLoadingRef && observerRef.current) { // Use the captured value
-        observerRef.current.unobserve(currentLoadingRef);
-      }
-      observerRef.current?.disconnect();
-    };
-  }, [hasNextPage, isFetchingNextPage, fetchNextPage]);
-
-  if (status === "pending") {
-    return <RecentListingsSkeleton />;
-  }
-
-  if (status === "error") {
-    return (
-      <div className="col-span-full text-center py-8 text-red-600">
-        Error: {error.message}
-      </div>
-    );
-  }
-
-  return (
-    <>
-      {data.pages.map((page, i) => (
-        <React.Fragment key={i}>
-          {page.map((listing: DisplayListingItem) => (
-            <Link key={listing.id} href={`/listings/${listing.id}`}>
-              <Card className="hover:shadow-md transition-shadow cursor-pointer overflow-hidden border-0">
-                <CardContent className="p-0">
-                  <div className="aspect-square bg-muted">
-                    <Image
-                      src={
-                        Array.isArray(listing.images)
-                          ? listing.images[0] || "/placeholder.svg"
-                          : listing.images || "/placeholder.svg"
-                      }
-                      alt={listing.title}
-                      width={200}
-                      height={200}
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div className="p-3">
-                    <h3 className="font-medium text-base mb-1 truncate">
-                      {listing.title}
-                    </h3>
-                    <p className="text-lg font-bold text-green-600 mb-1">
-                      Ksh {listing.price ?? "N/A"}
-                    </p>
-                    <p className="text-sm text-muted-foreground line-clamp-2 mb-1">
-                      {listing.description}
-                    </p>
-                    <div className="flex items-center gap-2 mb-1">
-                      <Badge variant="outline" className="text-xs">
-                        {listing.condition}
-                      </Badge>
-                    </div>
-                    <div className="flex items-center justify-between text-xs text-muted-foreground">
-                      <div className="flex items-center gap-1">
-                        <MapPin className="h-3 w-3" />
-                        {listing.location}
-                      </div>
-                      <div className="flex items-center gap-1">
-                        <div className="h-3 w-3 text-white" />
-                        {listing.views}
-                      </div>
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            </Link>
-          ))}
-        </React.Fragment>
-      ))}
-      <div ref={loadingRef} />
-      {isFetchingNextPage && <RecentListingsSkeleton />}
-    </>
-  );
-}
-
-export default function HomePage() {
-  const [searchQuery, setSearchQuery] = useState("");
-  const router = useRouter();
-
-  const handleSearch = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (searchQuery.trim()) {
-      router.push(`/listings?search=${encodeURIComponent(searchQuery)}`);
-    }
-  };
+export default async function HomePage() {
+  // Fetch initial listings on the server
+  const initialListings = await getRecentListings(1);
 
   return (
     <div className="min-h-screen bg-background">
@@ -171,23 +39,7 @@ export default function HomePage() {
             </p>
 
             {/* Search Bar */}
-            <div className="relative max-w-2xl mx-auto">
-              <form className="relative" onSubmit={handleSearch}>
-                <Search className="absolute left-4 top-1/2 h-5 w-5 -translate-y-1/2 text-gray-400" />
-                <Input
-                  placeholder="Search for items, services..."
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  className="pl-12 pr-4 py-6 text-base bg-white text-gray-900 border-0 h-12"
-                />
-                <Button
-                  size="sm"
-                  className="absolute right-2 top-1/2 -translate-y-1/2"
-                >
-                  Search
-                </Button>
-              </form>
-            </div>
+            <SearchBar />
           </div>
         </div>
       </section>
@@ -215,7 +67,7 @@ export default function HomePage() {
               }
             >
               <Suspense fallback={<RecentListingsSkeleton />}>
-                <RecentListings />
+                <RecentListings initialListings={initialListings} />
               </Suspense>
             </ErrorBoundary>
           </div>
